+++ changelog.md
@@ -9,4 +9,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2023-01-27
 ### Added
-- Runtime music generation
+- RandomTracksPlayer to play random tracks from a list of tracks